bugfix: ia_monture ordres trier numeriquement

+++ fonctions_sql/ia_monture.sql
@@ -166,7 +166,7 @@ begin
 
   -- -------------------------------------------------------------------------------------------------------------------
   -- monture à ordre ou mixte, reéalise un ordre, recherche de l'ordre à executer
-  select v into v_ordre from ( select json_array_elements(v_param) as v from perso where perso_cod=v_monstre ) as s order by v->>'ordre' limit 1 ;
+  select v into v_ordre from ( select json_array_elements(v_param) as v from perso where perso_cod=v_monstre ) as s order by (v->>'ordre')::integer limit 1 ;
   if not found  then
       -- on consomme quand même des PA, la monture ne garde pas de PA réserve pour plus tard !
       update perso set perso_pa = GREATEST(0, perso_pa - 4) where perso_cod=v_monstre ;
